backend/requirements.txt
# Core
fastapi==0.121.3
uvicorn==0.38.0
SQLAlchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11
python-multipart

# Validation & Settings
pydantic==2.12.4
pydantic-settings==2.6.1
python-dotenv==1.2.1
email-validator==2.3.0

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.32.3

# Dev / Test
pytest==9.0.1
black==24.8.0
isort==5.13.2
mypy==1.11.1
bandit==1.7.10

# Async stack
anyio==4.11.0
sniffio==1.3.1
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
httpx==0.24.1